OpenGL & ImGui OO organizing, yet with a strange error
