+++ opengl_IMGUI/imgui.ini
@@ -4,7 +4,7 @@ Size=400,400
 Collapsed=0
 
 [Window][Dear ImGui Demo]
-Pos=20,22
+Pos=112,18
 Size=456,680
 Collapsed=0
