+++ opengl_IMGUI/opengl_IMGUI.vcxproj
@@ -20,6 +20,8 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\glad.c" />
+    <ClCompile Include="src\graphicEngine.cpp" />
+    <ClCompile Include="src\gui.cpp" />
     <ClCompile Include="src\imgui.cpp" />
     <ClCompile Include="src\imgui_demo.cpp" />
     <ClCompile Include="src\imgui_draw.cpp" />
@@ -32,6 +34,8 @@
     <ClInclude Include="dependencies\glad.h" />
     <ClInclude Include="dependencies\glfw3.h" />
     <ClInclude Include="dependencies\glfw3native.h" />
+    <ClInclude Include="dependencies\graphicEngine.h" />
+    <ClInclude Include="dependencies\gui.h" />
     <ClInclude Include="dependencies\imconfig.h" />
     <ClInclude Include="dependencies\imgui.h" />
     <ClInclude Include="dependencies\imgui_impl_glfw.h" />
