+++ opengl_IMGUI/src/graphicEngine.cpp
@@ -1,0 +1,59 @@
+#include <stdio.h>
+#include <stdexcept>
+#include "graphicEngine.h"
+
+static void glfw_error_callback(int error, const char* description)
+{
+	fprintf(stderr, "Glfw Error %d: %s\n", error, description);
+}
+
+GraphicEngine::GraphicEngine()
+{
+	gl_init();
+}
+
+void GraphicEngine::gl_init()
+{
+	// Setup window
+	glfwSetErrorCallback(glfw_error_callback);
+	if (!glfwInit())
+		throw std::runtime_error("glfwInit() error");
+
+	// Decide GL+GLSL versions
+#if __APPLE__
+	// GL 3.2 + GLSL 150
+	glsl_version = "#version 150";
+	glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
+	glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 2);
+	glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);  // 3.2+ only
+	glfwWindowHint(GLFW_OPENGL_FORWARD_COMPAT, GL_TRUE);            // Required on Mac
+#else
+	// GL 3.0 + GLSL 130
+	glsl_version = "#version 130";
+	glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
+	glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 3);
+	//glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);  // 3.2+ only
+	//glfwWindowHint(GLFW_OPENGL_FORWARD_COMPAT, GL_TRUE);            // 3.0+ only
+#endif
+
+	// Create window with graphics context
+	window = glfwCreateWindow(1280, 720, "GLFWindow", NULL, NULL);
+	if (window == NULL)
+		throw std::runtime_error("glfwWindow creation error");
+
+	glfwMakeContextCurrent(window);
+	glfwSwapInterval(1); // Enable vsync
+
+	bool err = gladLoadGL() == 0;
+	if (err)
+	{
+		fprintf(stderr, "Failed to initialize OpenGL loader!\n");
+		throw std::runtime_error("OpenGL loader init error");
+	}
+}
+
+GraphicEngine::~GraphicEngine()
+{
+	glfwDestroyWindow(window);
+	glfwTerminate();
+}
